refactor(sort): create `sortByCreatedAt()`, remove `handleSorting`
- Add moment-timezone explicitly
- Move `util/date` to TypeScript
- Create `sortByCreatedAt()`, which handles sorting of entries by their
  `createdAt` date
- Use `sortByCreatedAt()` in the two places where `handleSorting` was
  previously used
- Remove `handleSorting`
