+++ client/package.json
@@ -17,6 +17,7 @@
     "framer-motion": "^4.1.17",
     "fuse.js": "^6.4.6",
     "moment": "2.29.1",
+    "moment-timezone": "^0.5.33",
     "prop-types": "^15.7.2",
     "query-string": "^7.0.1",
     "react": "^17.0.2",
