+++ client/src/pages/Post/AnswerSection/AnswerSection.component.jsx
@@ -1,20 +1,18 @@
-import React, { useState } from 'react'
 import { useQuery } from 'react-query'
 import Spinner from '../../../components/Spinner/Spinner.component'
-import handleSorting from '../../../services/handleSorting'
 import AnswerItem from './AnswerItem/AnswerItem.component'
 import './AnswerSection.styles.scss'
 import {
   getAnswersForPost,
   GET_ANSWERS_FOR_POST_QUERY_KEY,
 } from '../../../services/AnswerService'
+import { sortByCreatedAt } from '../../../util/date'
 
 const AnswerSection = ({ post }) => {
   const { isLoading, data: answers } = useQuery(
     [GET_ANSWERS_FOR_POST_QUERY_KEY, post.id],
     () => getAnswersForPost(post.id),
   )
-  const [sortType] = useState('Newest')
 
   return isLoading ? (
     <Spinner width="25px" height="25px" />
@@ -26,7 +24,7 @@ const AnswerSection = ({ post }) => {
         </div>
       )}
       {answers?.length > 0
-        ? answers?.sort(handleSorting(sortType)).map((answer) => (
+        ? answers?.sort(sortByCreatedAt).map((answer) => (
             <div key={answer.id} className="answers">
               <AnswerItem answer={answer} postId={post.id} />
             </div>
